+++ example/lib/routes/app_pages.dart
@@ -1,8 +1,6 @@
 // ignore_for_file: unused_import
 import 'package:get/get.dart';
 import 'package:example/routes/app_routes.dart';
-import 'package:example/modules/orders/pages/orders_page.dart';
-import 'package:example/modules/orders/bindings/orders_binding.dart';
 
 /// Route → Page + Binding mapping.
 abstract class AppPages {
@@ -16,6 +14,5 @@ abstract class AppPages {
     //   binding: HomeBinding(),
     //   transition: defaultTransition,
     // ),
-    GetPage(name: AppRoutes.orders, page: OrdersPage.new, binding: OrdersBinding(), transition: defaultTransition),
   ];
 }