+++ example/lib/routes/app_routes.dart
@@ -2,12 +2,12 @@
 abstract class AppRoutes {
   static const initial = _Routes.home;
   static const home = _Routes.home;
+  static const login = _Routes.login;
 
   // TODO: add routes here (ag module generates stubs in ROUTE_TODO.md)
-  static const orders = _Routes.orders;
 }
 
 abstract class _Routes {
   static const home = '/home';
-  static const orders = '/orders';
+  static const login = '/login';
 }